backend/apps/beaver_linter/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["nikita zigman <zigman.nikita@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0.2"



[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
types-toml = "^0.10.8.7"
testcontainers = {extras = ["mongodb"], version = "^3.7.1"}



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
